% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NicheNet Pipeline.R
\name{testFunction}
\alias{testFunction}
\title{this is my function}
\usage{
testFunction(a)
}
\arguments{
\item{number}{to print}
}
\value{
no return value
}
\description{
Theoretically it does something
I really hope it actually does
}
